34 create a way to spend money

+++ Realm/Assets/Scripts/GatherMoneyWinCondition.cs
@@ -1,6 +1,6 @@
 ﻿public class GatherMoneyWinCondition : WinCondition
 {
-    public decimal Goal { get; set; }
+    public float Goal { get; set; }
     
     public override WinConditionCheckResult Check(GameState gameState)
     {
